NOTICK: Rename configuration library code to avoid component confusion
