+++ applications/examples/demo-app/build.gradle
@@ -27,8 +27,8 @@ dependencies {
     implementation project(':libs:configuration:configuration-read')
 
     runtimeOnly project(":libs:messaging:kafka-messaging-impl")
-    runtimeOnly project(":libs:configuration:kafka-configuration-write-impl")
-    runtimeOnly project(":libs:configuration:kafka-configuration-read-impl")
+    runtimeOnly project(":libs:configuration:configuration-write-impl")
+    runtimeOnly project(":libs:configuration:configuration-read-impl")
     runtimeOnly project(":libs:kafka-utils:topic-admin-impl")
     runtimeOnly project(":libs:schema-registry:schema-registry-impl")
     runtimeOnly project(":libs:lifecycle:lifecycle-impl")
