+++ applications/tools/kafka-setup/build.gradle
@@ -28,7 +28,7 @@ dependencies {
 
     runtimeOnly project(':components:kafka-topic-admin')
     runtimeOnly project(':components:kafka-config-write')
-    runtimeOnly project(":libs:configuration:kafka-configuration-write-impl")
+    runtimeOnly project(":libs:configuration:configuration-write-impl")
     runtimeOnly project(":libs:kafka-utils:topic-admin-impl")
     runtimeOnly project(":libs:lifecycle:lifecycle-impl")
     runtimeOnly project(":libs:messaging:kafka-messaging-impl")
