+++ libs/configuration/configuration-read-impl/detekt-baseline.xml
