+++ libs/configuration/configuration-read-impl/src/main/resources/kafka.properties
