+++ libs/configuration/configuration-read-impl/src/test/resources/kafka.conf
