--- libs/configuration/configuration-read/src/main/kotlin/net/corda/libs/configuration/read/ConfigReadService.kt
@@ -1,12 +1,0 @@
-package net.corda.libs.configuration.read
-
-import net.corda.lifecycle.Lifecycle
-
-
-interface ConfigReadService : Lifecycle {
-    /**
-     * Register a callback for any configuration changes
-     * If the service is already running, you will receive a snapshot of all available configurations
-     */
-    fun registerCallback(configListener: ConfigListener): AutoCloseable
-}
