+++ libs/configuration/configuration-read/src/main/kotlin/net/corda/libs/configuration/read/factory/ConfigReaderFactory.kt
@@ -1,0 +1,13 @@
+package net.corda.libs.configuration.read.factory
+
+import com.typesafe.config.Config
+import net.corda.libs.configuration.read.ConfigReader
+
+interface ConfigReaderFactory {
+
+    /**
+     * Create an instance of the [ConfigReader]
+     * @param bootstrapConfig configuration object used to bootstrap the read service
+     */
+    fun createReader(bootstrapConfig: Config): ConfigReader
+}
