+++ libs/configuration/configuration-write-impl/detekt-baseline.xml
