+++ libs/configuration/configuration-write-impl/src/test/resources/config.conf
