--- libs/configuration/kafka-configuration-read-impl/build.gradle
@@ -1,23 +1,0 @@
-plugins {
-    id 'corda.common-library'
-    id 'com.r3.internal.gradle.plugins.r3Publish'
-}
-
-description 'Kafka Configuration Read API Impl'
-
-dependencies {
-    compileOnly 'org.osgi:osgi.annotation'
-    compileOnly 'org.osgi:org.osgi.service.component.annotations'
-
-    implementation platform("net.corda:corda-api:$cordaApiVersion")
-    implementation 'net.corda.kotlin:kotlin-stdlib-jdk8-osgi'
-    implementation "com.typesafe:config:$typeSafeConfigVersion"
-    implementation "net.corda:corda-base"
-    implementation "net.corda:avro-schema"
-    implementation project(':libs:configuration:configuration-read')
-    implementation project(":libs:lifecycle:lifecycle")
-    implementation project(":libs:messaging:messaging")
-
-    testImplementation "org.mockito:mockito-core:$mockitoVersion"
-    testImplementation "org.mockito.kotlin:mockito-kotlin:$mockitoKotlinVersion"
-}
