--- libs/configuration/kafka-configuration-write-impl/build.gradle
@@ -1,24 +1,0 @@
-plugins {
-    id 'corda.common-library'
-    id 'com.r3.internal.gradle.plugins.r3Publish'
-}
-
-description 'Kafka Configuration Write API Impl'
-
-
-dependencies {
-    compileOnly 'org.osgi:org.osgi.service.component.annotations'
-
-    implementation platform("net.corda:corda-api:$cordaApiVersion")
-    implementation 'net.corda.kotlin:kotlin-stdlib-jdk8-osgi'
-    implementation "net.corda:corda-base"
-    implementation "org.apache.servicemix.bundles:org.apache.servicemix.bundles.kafka-clients:$kafkaClientVersion"
-    implementation "com.typesafe:config:$typeSafeConfigVersion"
-    implementation "net.corda:avro-schema"
-
-    implementation project(":libs:configuration:configuration-write")
-    implementation project(":libs:messaging:messaging")
-
-    testImplementation "org.mockito:mockito-core:$mockitoVersion"
-    testImplementation "org.mockito.kotlin:mockito-kotlin:$mockitoKotlinVersion"
-}
