+++ libs/install/tests.bndrun
@@ -44,6 +44,7 @@ java: ${project.javaExecutable}
 	com.sun.activation.javax.activation;version='[1.2.0,1.2.1)',\
 	com.typesafe.config;version='[1.4.1,1.4.2)',\
 	javassist;version='[3.27.0,3.27.1)',\
+	javax.activation-api;version='[1.2.0,1.2.1)',\
 	javax.interceptor-api;version='[1.2.0,1.2.1)',\
 	javax.persistence-api;version='[2.2.0,2.2.1)',\
 	jaxb-api;version='[2.3.1,2.3.2)',\
