+++ settings.gradle
@@ -76,9 +76,9 @@ include 'components:examples:pubsub-sub'
 include 'components:examples:state-plus-event-sub'
 include 'libs:class-info'
 include 'libs:configuration:configuration-read'
+include 'libs:configuration:configuration-read-impl'
 include 'libs:configuration:configuration-write'
-include 'libs:configuration:kafka-configuration-read-impl'
-include 'libs:configuration:kafka-configuration-write-impl'
+include 'libs:configuration:configuration-write-impl'
 include 'libs:crypto:crypto-internal'
 include 'libs:crypto:crypto-impl'
 include 'libs:crypto:crypto-testkit'
